scripts/subida_historica_revanchas/requirements.txt
openpyxl==3.1.2
supabase==1.0.4
python-dotenv==1.0.0
